requirements.txt
numpy==1.26.4
opencv-python==4.10.0.84
pyopengl==3.1.0
scikit-image==0.24.0
scipy
tensorboard
smplx==0.1.28
spacepy==0.6.0
torchgeometry==0.1.2
tqdm
trimesh==4.5.1
pyrender==0.1.45